Addressing comments prt. 2
