+++ quipucords/api/common/util.py
@@ -353,3 +353,11 @@ def extract_tar_gz(file_like_obj):
                 return None
         file_data_list.append(file_data)
     return file_data_list
+
+def create_filename(report_type, file_ext, report_id):
+    """Create the filename."""
+    file_name = 'report_id_%s/%s_%s.%s' % (report_id,
+                                           report_type,
+                                           time.strftime('%Y%m%d%H%M%S'),
+                                           file_ext)
+    return file_name
