+++ quipucords/api/details_report/csv_renderer.py
@@ -10,13 +10,12 @@
 #
 """CSV renderer for reports."""
 
-from api.deployments_report.util import create_deployments_csv
 from api.details_report.util import create_details_csv
 
 from rest_framework import renderers
 
 
-class ReportCSVRenderer(renderers.BaseRenderer):
+class DetailsCSVRenderer(renderers.BaseRenderer):
     """Class to render report as CSV."""
 
     # pylint: disable=too-few-public-methods
@@ -29,9 +28,4 @@ class ReportCSVRenderer(renderers.BaseRenderer):
                renderer_context=None):
         """Render report as CSV."""
         # pylint: disable=arguments-differ,unused-argument,too-many-locals
-        report_type = report_dict.get('report_type')
-        if report_type == 'details':
-            return create_details_csv(report_dict)
-        if report_type == 'deployments':
-            return create_deployments_csv(report_dict)
-        return None
+        return create_details_csv(report_dict)
